+++ flythings/__init__.py
@@ -22,4 +22,5 @@ from flythings.client import search, send_observation, send_observations, login,
     find_series, send_record, register_action, register_action_for_series, start_action_listening, \
     stop_action_listening, ActionDataTypes, load_data_by_file, set_batch_enabled, get_observation_csv, \
     send_observations_csv, get_headers, api_get_request, send_alert, save_text_metadata, save_date_metadata, \
-    send_progress_action, send_prediction, send_predictions, search_prediction
+    send_progress_action, send_prediction, send_predictions, search_prediction, get_text_metadata, \
+    get_infrastructure_withmetadata, create_infrastructure, update_infrastructure